pyproject.toml
[tool.poetry]
name = "forecasting-ms"
version = "0.1.0"
description = ""
authors = ["Xavier nebotxavi@gmail.com"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12"
fastapi = ">=0.115.13,<0.116.0"
uvicorn = ">=0.34.3,<0.35.0"
httpx = ">=0.28.1,<0.29.0"
pandas = ">=2.3.0,<3.0.0"
prophet = ">=1.1.7,<2.0.0"
pydantic-settings = ">=2.10.1,<3.0.0"
nats-py = ">=2.10.0,<3.0.0"
pytest-asyncio = ">=1.0.0,<2.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
poethepoet = "^0.35.0"

[tool.poe.tasks]
start = "uvicorn app.drivers.main:app --reload"
